requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.6.0
httpx[http2]>=0.27.0
python-dotenv>=1.0.0
pydub>=0.25.1
chromadb>=0.4.24
google-genai>=1.0.0
groq>=0.5.0
openai>=1.12.0
celery>=5.3.0
redis>=5.0.0
